requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
imbalanced-learn>=0.11
shap>=0.43
matplotlib>=3.7
openpyxl>=3.1
pyyaml>=6.0
